refactor(harness): extract pi Elixir adapter to Cli.Harness.Pi
Step 1c of multi-harness support (sessions#50). Pure refactor — 154/154
bats tests and 4 doctests + 73 elixir tests still green.

Moves all pi-specific knowledge out of Cli:
- Command construction (pi CLI flags, model qualification, shell script)
- Stream protocol (message_update/assistantMessageEvent parsing,
  toolcall_start/delta/end, agent_end usage extraction)
- Tool input argument formatting (bash/edit/pattern/path/url/prompt/todos)
- Partial-JSON text recovery regex

Cli keeps orchestration (parse_args, validate_args, port spawning,
stream_output loop) and genuinely harness-agnostic helpers
(check_abort_signal, text_beyond_flushed, print_usage_summary).

Public Cli.process_line/2, Cli.format_tool_input/1,
Cli.extract_partial_text/1, and Cli.flush_partial_buffer/1 remain as
delegating wrappers so existing tests and callers don't break. Step 2
will turn them into dispatchers.
