+++ cli/lib/cli.ex
@@ -1,16 +1,18 @@
 defmodule Cli do
   @moduledoc """
-  Session execution engine — runs pi with streaming output, timeout,
-  ABORT detection, and usage reporting.
+  Session execution engine — runs an agent harness with streaming output,
+  timeout, ABORT detection, and usage reporting.
 
   Identity-agnostic: receives a system prompt file, doesn't know or care
   what's in it. Prompt composition (identity, passphrase, context) is the
   caller's responsibility.
+
+  Today this engine only knows pi. All pi-specific logic lives in
+  `Cli.Harness.Pi`; this module keeps the CLI surface, port spawning,
+  stream loop, and harness-agnostic helpers (abort detection, usage
+  summary). See sessions#50 for the multi-harness plan.
   """
 
-  @default_model "claude-opus-4-6"
-  @truncate_edit_limit 60
-  @truncate_prompt_limit 100
   @buffer_flush_timeout_ms 100
   @timeout_exit_code 124
 
@@ -34,7 +36,7 @@ defmodule Cli do
   defp run_with_opts(opts, rest) do
     message = Enum.join(rest, " ")
     timeout = opts[:timeout]
-    model = opts[:model] || @default_model
+    model = opts[:model] || Cli.Harness.Pi.default_model()
     cwd = opts[:cwd]
     session = opts[:session]
 
@@ -120,6 +122,8 @@ defmodule Cli do
   end
 
   defp print_help do
+    default_model = Cli.Harness.Pi.default_model()
+
     IO.puts("""
     Usage: sessions run --system-prompt-file <path> [options] <message>
 
@@ -130,7 +134,7 @@ defmodule Cli do
 
     Options:
       --timeout <seconds>          Maximum runtime in seconds (default: no timeout)
-      --model <model>              Model to use (default: #{@default_model})
+      --model <model>              Model to use (default: #{default_model})
       --cwd <path>                 Working directory for pi
       --session <path>             Session file for conversation continuity
       --no-extensions              Disable pi extensions
@@ -146,40 +150,19 @@ defmodule Cli do
   end
 
   defp run_agent(message, system_prompt_file, timeout, model, cwd, session, pi_opts) do
-    # Ensure model includes a provider prefix (e.g. "anthropic/claude-opus-4-6").
-    qualified_model =
-      if String.contains?(model, "/"), do: model, else: "anthropic/#{model}"
+    {shell_script, positional_args} =
+      Cli.Harness.Pi.build_command(
+        message,
+        model,
+        system_prompt_file,
+        session,
+        timeout,
+        pi_opts[:extensions],
+        pi_opts[:skills],
+        pi_opts[:prompt_templates]
+      )
 
-    # Build pi flags. User-controlled strings use positional args to avoid shell injection.
-    # $1=message, $2=model, $3=system_prompt_file, $4=session (if provided)
-    session_flag =
-      if session, do: ~s( --session "$4"), else: " --no-session"
-
-    pi_flags =
-      [
-        ~s( --append-system-prompt "$3"),
-        ~s( --model "$2"),
-        " --mode json",
-        session_flag,
-        if(pi_opts[:extensions], do: "", else: " --no-extensions"),
-        if(pi_opts[:skills], do: "", else: " --no-skills"),
-        if(pi_opts[:prompt_templates], do: "", else: " --no-prompt-templates")
-      ]
-      |> Enum.join("")
-
-    pi_cmd = ~s(pi -p "$1"#{pi_flags})
-
-    # echo | pipes empty stdin so pi doesn't block waiting for terminal input.
-    # pi -p (print mode) reads stdin before processing the message.
-    shell_script =
-      if timeout do
-        "echo | timeout #{timeout} #{pi_cmd}"
-      else
-        "echo | #{pi_cmd}"
-      end
-
-    args = ["-c", shell_script, "--", message, qualified_model, system_prompt_file]
-    args = if session, do: args ++ [session], else: args
+    args = ["-c", shell_script, "--"] ++ positional_args
 
     port_opts = [:binary, :exit_status, :stderr_to_stdout, {:args, args}]
     port_opts = if cwd, do: [{:cd, cwd} | port_opts], else: port_opts
@@ -272,23 +255,21 @@ defmodule Cli do
     end
   end
 
+  # --- Delegating wrappers ---
+  #
+  # These exist so existing callers, tests, and doctests don't break
+  # during the multi-harness refactor. Step 2 will make `process_line`
+  # (and friends) dispatch on a harness resolved from session context.
+
   @doc """
   Extract text from incomplete JSON lines in the buffer.
   Returns the unescaped text, or empty string if none found.
+
+  Delegates to the pi harness adapter today; will dispatch on harness
+  in step 2.
   """
   @spec extract_partial_text(String.t()) :: String.t()
-  def extract_partial_text(partial) do
-    case Regex.run(~r/"(?:text|delta)"\s*:\s*"((?:[^"\\]|\\.)*)$/, partial) do
-      [_, text] ->
-        case Jason.decode("\"#{text}\"") do
-          {:ok, unescaped} -> unescaped
-          {:error, _} -> ""
-        end
-
-      nil ->
-        ""
-    end
-  end
+  defdelegate extract_partial_text(partial), to: Cli.Harness.Pi
 
   @doc """
   Flush incomplete JSON lines from the buffer to stdout.
@@ -301,168 +282,23 @@ defmodule Cli do
     end
   end
 
-  @typep stream_state :: %{
-           tool_input: String.t(),
-           buffer: String.t(),
-           usage: map() | nil,
-           abort_seen: boolean(),
-           recent_text: String.t(),
-           flushed_chars: non_neg_integer(),
-           had_newline_before_window: boolean()
-         }
-
   @doc """
-  Detect [[ABORT]] signal on its own line, handling chunk boundaries.
-  Returns {abort_seen, recent_text, had_newline_before_window}.
+  Format tool input map into a human-readable string for display.
+  Returns nil for unrecognized input formats.
   """
-  @spec check_abort_signal(String.t(), stream_state()) ::
-          {boolean(), String.t(), boolean()}
-  def check_abort_signal(text, state) do
-    combined = state.recent_text <> text
-    combined_len = String.length(combined)
-
-    trimmed_len = max(0, combined_len - 20)
-    trimmed_portion = String.slice(combined, 0, trimmed_len)
-
-    had_newline_before_window =
-      if String.contains?(trimmed_portion, "\n"),
-        do: true,
-        else: state.had_newline_before_window
-
-    text_to_check =
-      if had_newline_before_window, do: "\n" <> combined, else: combined
-
-    abort_seen =
-      state.abort_seen ||
-        Regex.match?(~r/(?:^|\n)\[\[ABORT\]\](?:\n|$)/, text_to_check)
-
-    recent_text = String.slice(combined, -20, 20)
-
-    {abort_seen, recent_text, had_newline_before_window}
-  end
+  @spec format_tool_input(map()) :: String.t() | nil
+  defdelegate format_tool_input(input), to: Cli.Harness.Pi
 
   @doc false
-  @spec process_line(String.t(), stream_state()) :: stream_state()
+  @spec process_line(String.t(), map()) :: map()
   def process_line(line, state) do
-    case Jason.decode(line) do
-      {:ok,
-       %{
-         "type" => "message_update",
-         "assistantMessageEvent" => %{"type" => "text_delta", "delta" => text}
-       }} ->
-        handle_text_delta(text, state)
-
-      {:ok,
-       %{
-         "type" => "message_update",
-         "assistantMessageEvent" => %{"type" => "toolcall_start"} = event
-       }} ->
-        name = extract_tool_name_from_event(event)
-        IO.puts("\n[TOOL] #{name}")
-        %{state | tool_input: ""}
-
-      {:ok,
-       %{
-         "type" => "message_update",
-         "assistantMessageEvent" => %{"type" => "toolcall_delta", "delta" => json}
-       }}
-      when json != "" ->
-        %{state | tool_input: state.tool_input <> json}
-
-      {:ok,
-       %{
-         "type" => "message_update",
-         "assistantMessageEvent" => %{"type" => "toolcall_end", "toolCall" => tool_call}
-       }} ->
-        handle_tool_call_end(tool_call, state)
-
-      {:ok, %{"type" => "agent_end", "messages" => messages}} ->
-        handle_agent_end(messages, state)
-
-      _ ->
-        state
-    end
+    Cli.Harness.Pi.process_line(line, state,
+      check_abort: &__MODULE__.check_abort_signal/2,
+      text_beyond_flushed: &__MODULE__.text_beyond_flushed/2
+    )
   end
 
-  defp extract_tool_name_from_event(%{
-         "contentIndex" => idx,
-         "partial" => %{"content" => content}
-       }) do
-    case Enum.at(content, idx) do
-      %{"name" => name} -> name
-      _ -> "unknown"
-    end
-  end
-
-  defp extract_tool_name_from_event(_), do: "unknown"
-
-  defp handle_text_delta(text, state) do
-    text_to_write = text_beyond_flushed(text, state.flushed_chars)
-    maybe_write_text(text_to_write)
-
-    {abort_seen, recent_text, had_newline} = check_abort_signal(text, state)
-
-    %{
-      state
-      | abort_seen: abort_seen,
-        recent_text: recent_text,
-        flushed_chars: 0,
-        had_newline_before_window: had_newline
-    }
-  end
-
-  defp maybe_write_text(""), do: :ok
-  defp maybe_write_text(text), do: IO.write(text)
-
-  defp handle_tool_call_end(tool_call, state) do
-    case Map.get(tool_call, "arguments") do
-      nil -> :ok
-      args -> print_tool_input(args)
-    end
-
-    %{state | tool_input: ""}
-  end
-
-  defp handle_agent_end(messages, state) do
-    assistant_msgs =
-      Enum.filter(messages, fn msg ->
-        msg["role"] == "assistant" && msg["usage"] != nil
-      end)
-
-    totals =
-      Enum.reduce(
-        assistant_msgs,
-        %{input: 0, output: 0, cache_read: 0, cache_write: 0, cost: 0.0},
-        fn msg, acc ->
-          u = msg["usage"]
-          cost = get_in(u, ["cost", "total"]) || 0.0
-
-          %{
-            input: acc.input + (u["input"] || 0),
-            output: acc.output + (u["output"] || 0),
-            cache_read: acc.cache_read + (u["cacheRead"] || 0),
-            cache_write: acc.cache_write + (u["cacheWrite"] || 0),
-            cost: acc.cost + cost
-          }
-        end
-      )
-
-    %{
-      state
-      | usage: %{
-          cost_usd: totals.cost,
-          duration_ms: nil,
-          num_turns: length(assistant_msgs),
-          usage: %{
-            "input_tokens" => totals.input,
-            "output_tokens" => totals.output,
-            "cache_read_input_tokens" => totals.cache_read,
-            "cache_creation_input_tokens" => totals.cache_write
-          },
-          model_usage: nil
-        }
-    }
-  end
+  # --- Harness-agnostic helpers ---
 
   @doc """
   Returns the portion of `text` beyond already-flushed characters.
@@ -491,6 +327,45 @@ defmodule Cli do
     else
       ""
     end
+  end
+
+  @typep abort_state :: %{
+           optional(:abort_seen) => boolean(),
+           optional(:recent_text) => String.t(),
+           optional(:had_newline_before_window) => boolean()
+         }
+
+  @doc """
+  Detect [[ABORT]] signal on its own line, handling chunk boundaries.
+  Returns {abort_seen, recent_text, had_newline_before_window}.
+
+  Harness-agnostic — operates on raw text, not on any specific event
+  schema.
+  """
+  @spec check_abort_signal(String.t(), abort_state()) ::
+          {boolean(), String.t(), boolean()}
+  def check_abort_signal(text, state) do
+    combined = state.recent_text <> text
+    combined_len = String.length(combined)
+
+    trimmed_len = max(0, combined_len - 20)
+    trimmed_portion = String.slice(combined, 0, trimmed_len)
+
+    had_newline_before_window =
+      if String.contains?(trimmed_portion, "\n"),
+        do: true,
+        else: state.had_newline_before_window
+
+    text_to_check =
+      if had_newline_before_window, do: "\n" <> combined, else: combined
+
+    abort_seen =
+      state.abort_seen ||
+        Regex.match?(~r/(?:^|\n)\[\[ABORT\]\](?:\n|$)/, text_to_check)
+
+    recent_text = String.slice(combined, -20, 20)
+
+    {abort_seen, recent_text, had_newline_before_window}
   end
 
   defp print_usage_summary(%{usage: nil}), do: :ok
@@ -522,111 +397,6 @@ defmodule Cli do
       if cache_read > 0 or cache_create > 0 do
         IO.puts("  Cache: #{cache_read} read, #{cache_create} created")
       end
-    end
-  end
-
-  defp print_tool_input(input) do
-    case format_tool_input(input) do
-      nil -> :ok
-      output -> IO.puts(output)
-    end
-  end
-
-  @doc """
-  Formats tool input map into a human-readable string for display.
-  Returns nil for unrecognized input formats.
-  """
-  @spec format_tool_input(map()) :: String.t() | nil
-  def format_tool_input(%{"command" => cmd}) do
-    "  $ #{cmd}"
-  end
-
-  def format_tool_input(%{"path" => path, "edits" => [first | _]}) when is_map(first) do
-    old =
-      first
-      |> Map.get("oldText", "")
-      |> truncate(@truncate_edit_limit)
-      |> String.replace("\n", "\\n")
-
-    new =
-      first
-      |> Map.get("newText", "")
-      |> truncate(@truncate_edit_limit)
-      |> String.replace("\n", "\\n")
-
-    "  #{path}\n  - #{old}\n  + #{new}"
-  end
-
-  def format_tool_input(%{"pattern" => pattern} = input) do
-    case Map.get(input, "path") do
-      nil -> "  pattern: #{pattern}"
-      path -> "  #{path}\n  pattern: #{pattern}"
-    end
-  end
-
-  def format_tool_input(%{"path" => path}) do
-    "  -> #{path}"
-  end
-
-  def format_tool_input(%{"url" => url, "prompt" => prompt} = input) do
-    prompt_preview = truncate(prompt, @truncate_prompt_limit)
-
-    case Map.get(input, "description") do
-      desc when desc in [nil, ""] -> "  url: #{url}\n  prompt: #{prompt_preview}"
-      desc -> "  #{desc}\n  url: #{url}\n  prompt: #{prompt_preview}"
-    end
-  end
-
-  def format_tool_input(%{"prompt" => prompt} = input) do
-    prompt_preview = truncate(prompt, @truncate_prompt_limit)
-
-    case Map.get(input, "description") do
-      desc when desc in [nil, ""] -> "  prompt: #{prompt_preview}"
-      desc -> "  #{desc}\n  prompt: #{prompt_preview}"
-    end
-  end
-
-  def format_tool_input(%{"todos" => todos}) when is_list(todos) do
-    count = length(todos)
-    first = List.first(todos)
-
-    preview =
-      case first do
-        %{"content" => content} when is_binary(content) -> ": #{truncate(content, 50)}"
-        _ -> ""
-      end
-
-    "  #{count} todo(s)#{preview}"
-  end
-
-  def format_tool_input(%{"query" => query}) do
-    "  search: #{truncate(query, 80)}"
-  end
-
-  def format_tool_input(%{"operation" => op, "filePath" => path, "line" => line}) do
-    "  #{op} at #{path}:#{line}"
-  end
-
-  def format_tool_input(%{"skill" => skill} = input) do
-    case Map.get(input, "args") do
-      nil -> "  skill: #{skill}"
-      args -> "  skill: #{skill} #{truncate(args, 50)}"
-    end
-  end
-
-  def format_tool_input(%{"shell_id" => id}) do
-    "  shell: #{id}"
-  end
-
-  def format_tool_input(_), do: nil
-
-  defp truncate(nil, _limit), do: ""
-
-  defp truncate(string, limit) do
-    if String.length(string) > limit do
-      String.slice(string, 0, limit) <> "..."
-    else
-      string
     end
   end
 end
